added auth for front and backend

+++ frontend/src/components/WorkoutDetails.jsx
@@ -1,15 +1,22 @@
 import { useWorkoutsContext } from '../hooks/useWorkoutsContext'
+import { useAuthContext } from '../hooks/useAuthContext'
 
 //date fns
-import formatDistaneToNow  from 'date-fns/formatDistanceToNow'
 import formatDistanceToNow from 'date-fns/formatDistanceToNow'
 
 const WorkoutDetails = ({ workout }) => {
   const { dispatch } = useWorkoutsContext()
+  const { user } = useAuthContext()
 
   const handleClick = async () => {
+    if (!user){
+      return
+    }
     const response = await fetch('/api/workouts/'+ workout._id, {
-      method: 'DELETE'
+      method: 'DELETE',
+      headers: {
+        'Authorization': `Bearer ${user.token}`
+      }
     })
     const json = await response.json()
